Refactory command in spark sql
